DFRPlayerTestFullClip v1.0 created
This program illustrates what is needed in order to play a clip on the mini MP3 player and then make sure that the first clip is complete before playing the nexty clip.  It accounts for the MP3 player (apparently) returning two "complete" status' after playing a clip.  An explanation of this is behavior is contained in the comments at the top of this program.