Busy Pin Test program
This program uses the BUSY pin in the MP3 Player (connected to Photon D0) to determine when the Player is not busy (when a clip is complete).  I found that it is necessary to provide a delay of about 1 second after commanding a new clip before the BUSY line can be tested.  This program implements the 1 second timeout in a non-blocking manner.