test program to read analog and control servo
This test program allows the user (via the Particle console) to set the mini MP3 player volume and clip number and play the selected clip.  It samples the analog processing circuit envelope output every 10 ms, averages the samples, and controls the mouth servo.  It also records the max and min values found for the clip, so that the scaling values can be adjusted accordingly on a subsequent run.